refactor(cli): migrate runtime-recovery.js to TypeScript
Convert bin/lib/runtime-recovery.js (84 lines) to src/lib/runtime-recovery.ts
with typed StateClassification interface.

5 pure classifiers/parsers + 1 I/O function (getRecoveryCommand).
Co-locates tests with additional edge case coverage.

617 CLI tests pass. Coverage ratchet passes.

Relates to #924 (shell consolidation).
